refactor: drastically simplify our model types
We didn't need to handle parsing both strings and ints for these values
from the API, so we can drop our custom types and all of the various
casting/conversion calls we were doing.

+++ internal/fsutil/fsutil.go
@@ -8,7 +8,7 @@ import (
 )
 
 // FileExistsAndMatchesSize checks if a file exists and has the expected size.
-func FileExistsAndMatchesSize(filePath string, expectedSize uint64) bool {
+func FileExistsAndMatchesSize(filePath string, expectedSize int64) bool {
 	fileInfo, err := os.Stat(filePath)
 	if err != nil {
 		if errors.Is(err, fs.ErrNotExist) {
@@ -18,7 +18,7 @@ func FileExistsAndMatchesSize(filePath string, expectedSize uint64) bool {
 		return false
 	}
 
-	return uint64(fileInfo.Size()) == expectedSize
+	return fileInfo.Size() == expectedSize
 }
 
 // ShortenPath replaces the home directory path with ~
